+++ spec/features/staff/users_can_create_a_programme_level_activity_spec.rb
@@ -17,6 +17,193 @@ RSpec.feature "Users can create a programme activity" do
       expect(page).to have_content t("action.programme.create.success")
     end
 
+    context "validations" do
+      scenario "validation errors work as expected" do
+        parent = create(:fund_activity, organisation: user.organisation)
+        identifier = "foo"
+
+        visit activities_path
+        click_on t("page_content.organisation.button.create_activity")
+
+        # Don't provide a level
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.level.blank")
+
+        choose "Programme"
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("form.legend.activity.parent")
+
+        # Don't provide a parent
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.parent.blank")
+
+        choose parent.title
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("form.label.activity.delivery_partner_identifier")
+
+        # Don't provide an identifier
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.delivery_partner_identifier.blank")
+
+        fill_in "activity[delivery_partner_identifier]", with: identifier
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("form.label.activity.roda_identifier_fragment", level: "programme")
+
+        # Provide an invalid identifier
+        fill_in "activity[roda_identifier_fragment]", with: "!!!"
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.roda_identifier_fragment.invalid_characters")
+
+        fill_in "activity[roda_identifier_fragment]", with: identifier
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content custom_capitalisation(t("form.legend.activity.purpose", level: "programme (level B)"))
+        expect(page).to have_content t("form.hint.activity.title", level: "programme (level B)")
+
+        # Don't provide a title and description
+        click_button t("form.button.activity.submit")
+
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.title.blank")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.description.blank")
+
+        fill_in "activity[title]", with: Faker::Lorem.word
+        fill_in "activity[description]", with: Faker::Lorem.paragraph
+        click_button t("form.button.activity.submit")
+
+        expect(page).to have_content t("form.legend.activity.sector_category", level: "programme (level B)")
+
+        # Don't provide a sector category
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.sector_category.blank")
+
+        choose "Basic Education"
+        click_button t("form.button.activity.submit")
+
+        expect(page).to have_content t("form.legend.activity.sector", sector_category: "Basic Education", level: "programme (level B)")
+        # Don't provide a sector
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.sector.blank")
+
+        choose "Primary education"
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("form.legend.activity.programme_status", level: "programme (level B)")
+
+        # Don't provide a programme status
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content "can't be blank"
+
+        choose("activity[programme_status]", option: "07")
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("page_title.activity_form.show.dates", level: "programme (level B)")
+
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.dates")
+
+        # Dates cannot contain only a zero
+        fill_in "activity[planned_start_date(3i)]", with: 1
+        fill_in "activity[planned_start_date(2i)]", with: 0
+        fill_in "activity[planned_start_date(1i)]", with: 2010
+        fill_in "activity[planned_end_date(3i)]", with: 0
+        fill_in "activity[planned_end_date(2i)]", with: 12
+        fill_in "activity[planned_end_date(1i)]", with: 2010
+        click_button t("form.button.activity.submit")
+
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.dates")
+
+        fill_in "activity[planned_start_date(3i)]", with: 1
+        fill_in "activity[planned_start_date(2i)]", with: 12
+        fill_in "activity[planned_start_date(1i)]", with: 2010
+        fill_in "activity[planned_end_date(3i)]", with: 1
+        fill_in "activity[planned_end_date(2i)]", with: 12
+        fill_in "activity[planned_end_date(1i)]", with: 2010
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("form.legend.activity.geography")
+
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.geography.blank")
+
+        choose "Region"
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("form.label.activity.recipient_region")
+
+        # region has the default value already selected
+        click_button t("form.button.activity.submit")
+
+        expect(page).to have_content t("form.label.activity.intended_beneficiaries")
+
+        # Don't select any intended beneficiaries
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.intended_beneficiaries.blank")
+
+        check "Haiti"
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("form.label.activity.gdi")
+
+        # Don't select a GDI
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.gdi.blank")
+
+        choose "GDI not applicable"
+        click_button t("form.button.activity.submit")
+        expect(page). to have_content t("form.label.activity.collaboration_type")
+
+        # Collaboration_type has a pre-selected option
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("form.label.activity.flow")
+
+        # Flow has a default and can't be set to blank so we skip
+        select "ODA", from: "activity[flow]"
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("form.legend.activity.aid_type")
+
+        # Don't select an aid type
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.aid_type.blank")
+
+        choose("activity[aid_type]", option: "B02")
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("form.legend.activity.fstc_applies")
+
+        # Don't choose if fstc applies or not
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("activerecord.errors.models.activity.attributes.fstc_applies.inclusion")
+
+        choose("activity[fstc_applies]", option: true)
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content t("form.legend.activity.oda_eligibility")
+
+        # oda_eligibility has the default value already selected
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content Activity.find_by(delivery_partner_identifier: identifier).title
+      end
+
+      scenario "failing to select a country shows an error message" do
+        visit activities_path
+        click_on(t("page_content.organisation.button.create_activity"))
+
+        choose custom_capitalisation(t("page_content.activity.level.fund"))
+        click_button t("form.button.activity.submit")
+        fill_in "activity[delivery_partner_identifier]", with: "no-country-selected"
+        click_button t("form.button.activity.submit")
+        fill_in "activity[roda_identifier_fragment]", with: "roda-identifier"
+        click_button t("form.button.activity.submit")
+        fill_in "activity[title]", with: "My title"
+        fill_in "activity[description]", with: "My description"
+        click_button t("form.button.activity.submit")
+        choose "Basic Education"
+        click_button t("form.button.activity.submit")
+        choose "School feeding"
+        click_button t("form.button.activity.submit")
+        fill_in "activity[planned_start_date(3i)]", with: "01"
+        fill_in "activity[planned_start_date(2i)]", with: "01"
+        fill_in "activity[planned_start_date(1i)]", with: "2020"
+        click_button t("form.button.activity.submit")
+        choose "Country"
+        click_button t("form.button.activity.submit")
+        click_button t("form.button.activity.submit")
+        expect(page).to have_content "Recipient country can't be blank"
+      end
+    end
+
     scenario "the activity has the appropriate funding organisation defaults" do
       fund = create(:activity, level: :fund, organisation: user.organisation)
       identifier = "a-programme-has-a-funding-organisation"