+++ spec/features/staff/users_can_edit_an_activity_spec.rb
@@ -567,6 +567,15 @@ def assert_all_edit_links_go_to_the_correct_form_step(activity:)
   click_on(t("default.link.back"))
   click_on t("tabs.activity.details")
 
+  within(".fstc_applies") do
+    click_on(t("default.link.edit"))
+    expect(page).to have_current_path(
+      activity_step_path(activity, :fstc_applies)
+    )
+  end
+  click_on(t("default.link.back"))
+  click_on t("tabs.activity.details")
+
   within(".oda_eligibility") do
     click_on(t("default.link.edit"))
     expect(page).to have_current_path(