+++ spec/features/staff/users_can_view_programme_level_activities_spec.rb
@@ -1,4 +1,4 @@
-RSpec.feature "Users can view programme level activites" do
+RSpec.feature "Users can view programme level activities" do
   context "when the user belongs to BEIS" do
     let(:user) { create(:beis_user) }